requirements.txt
astropy==5.3.3
google-genai==1.41.0
gwosc==0.7.1
gwpy==3.0.8
matplotlib==3.5.3
numpy==1.23.5
openai==2.1.0
openpyxl==3.1.5
pandas==2.3.3
protobuf==5.29.5
python-dotenv==1.1.1
scipy==1.10.0
streamlit==1.37.0
